requirements.txt
pymongo==3.8.0
pytest==4.6.3
pytest-cov==2.7.1
python-decouple==3.1
